requirements.txt
matplotlib
torch
torchmetrics[image]
tensorboard
tqdm
torchvision
gdown
pyyaml
torchsummary